Change tests to fail by raising the original exception.
Previously, if a test could fail because of an unexpected exception,
the exception was caught, and fail() was called with the exception's
message.

In some cases, the exception message was less informative than was
desirable; the traceback was preferable.  So, re-plumb the unit
tests so that exceptions are bubbled back to the top level of the
test, and into the Test annotation class.  That causes JUnit to
report the original exception directly.