Fix a bug in text field action commands.
The default "ActionCommand" for a text field is the user-typed text.
Since button actions and text field actions are processed by the
same method, if the contents of a text field were updated to be the
same as the name of a button, the button's action could be
performed.  That was kind of bad...

This changes things so that all the text fields have an action
command the same as the field's name.