Prevent accidentally throwing away account data changes.
Previously, if you selected an account from the list, any unsaved
changes to the account data would be discarded to load the selected
account.  This was bad, since you could wind up throwing away a
password change, or other account data that needed to get saved.

This enhances handling with these features:
  * There is a new "Revert" button option that explicitly discards
    unsaved changes.
  * The account list is disabled when there are unsaved changes.
    This prevents discarding changes by changing accounts.
  * If there are unsaved changes, the only available options are
    "Revert" or "Add"/"Update".
  * Copying a generated password isn't allowed if there are unsaved
    field changes.  (This may be wrong...)