Make the password member data of AccountStoreSupport.
The AccountStoreSupport test superclass contained a hard-coded
password string that was used for all operations requiring one.
However, for the various TestAccountStoreV? test classes the
password is determined by the data file associated with the test,
and shouldn't be hard-coded in any other file.

So, make the password member data, just like the test data
representing the file contents.  Also, update the comments in
create_test_data to reflect this.