Refactor temporary file handling in unit tests.
This shifts the method factoring in AccountStoreSupport for reading
and writing temporary AccountStore files.  The previous handling had
a bug that it didn't clean up the files when it was done with them.
The factoring also pushed temp file handling up the call stack,
forcing it to be replicated.

Refactor so that temporary file handling is isolated in one place,
and fix the bug.