Update the unit tests not to depend on Account.equals().
Certain unit tests operated by asserting that two different Account
objects matched according to their equals() method.  With the
upcoming V2 file format changes, the matching needs to be more
generous, since Account objects will have hidden fields that
shouldn't matter to the unit tests.

This changes the tests to create an AccountTestData object that
wraps an Account, and can perform a comparison that checks only
the fields that matter for testing purposes.