Update Account to support reading account history.
This adds a new function for indexed access to an account's
history of updates.  This change brings with it various refactoring
changes:
  * To facilitate indexed access, the update history is now a List,
    rather than a SortedSet.
  * A new abstract class AccountData is created; Account is a
    subclass of this class, as is the value returned by the new
    update history access method.

Additionally, stuff is renamed as necessary for all the changes to
hold together, and there's more javadoc to explain it.

Finally, there's more sanity checking in readAccount().  Errors in
account data are signaled by throwing an exception, and a broken
assertion is replaced with a fixed sanity test that throws an
exception.