Update Save As dialog to have a confirmation password.
This updates AccountFileDialog so that the SAVE_DIALOG version of
the dialog includes two password fields.  Saving the file requires
the two passwords to match.

This also adds a small standalone program that allows testing the
file dialog behaviors by popping up dialogs in sequence.

The new widget is still imperfect:  The SAVE_DIALOG doesn't handle
keyboard focus the way I want:  Requests to direct focus to one or
the other of the password fields tend to get ignored.  Instead, the
keyboard focus tends to move to the file list at awkward times, and
getting focus out of that window requires a mouse click (boo!).