Improve coverage of file reading tests.
The tests for reading hard-coded sample files didn't assert that every
expected account record was successfully read.  In some cases, it was
possible that the file could be read incompletely and yet the tests
could still pass.

This change adds an assertion that the size of the account store read
from the file matches the expected size.