Use "editable" instead of "enabled" in AccountDataPanel
When the account data panel is displaying historical account
credentials, it's not allowed to change the entry.  Previously, that
was enforced by using `setEnable(false)` on the various text fields.
But that made it impossible to copy the text in the fields, which wasn't
fully ideal.

This changes the usage to `setEditable(false)` instead, which allows
copy/paste of the historical info, but not changing it.