Fix build.xml to work in all my environments.
To accommodate difficutlties with ant and CLASSPATH on my macOS laptop,
build.xml was depending on ~/lib to contain the JUnit JAR files, and any
other dependencies.  This didn't work in the Linux environment on
cheshire.

I've since figured out how to fix macOS properly, so update to go back
to depending on CLASSPATH instead, since it's slightly more flexible
across environments.